Revised the macros with new DSL
The goal is a clearer and easier to use macro fro creating vertex and
edge objects.

the new version has more robustness:

 Sections can be left out (massmatrix, prep, etc..)
 Massmatrix will automatically suit an DAE if you provide algebraic
states (those without a derivative partner)
 Sections can be in any order
 {Static,ODE}{Vertex,Edge} are selected automatically from the given
state vector description

The macro code itself is currently in two large monolithic functions
that should be easier to read with less indirection. A rewrite to remove
the (majority) code duplication should be undertaken.

A demonstration example is also included.

A description of the DSL: An edge or vertex is created with the

`@ Vertex` or `@ Edge` macros.

Following the macro, you define a number of sections describing the
dynamics. Those sections, in any order, are:

```
 @ prepare   - optional. Tasks done before creating the function
               (parameter scaling, @ asserts, etc..)
 @ massmatrix - optional. The mass matrix of the node or edge. If none
                is provided, it will use an identity matrix for all
                differential states, and 0 for all algebraic states.
 @ states  - required. A list of states, one per line. Pairs are made
             into differential variables, singles are made into
             algebraic variables.
 @ dynamics - required. The dynamic equations of the node or edge, in
              terms of the provided state vector and parameters.
```

An ODE{Vertex,Edge} is created if there is a single differential pair. A
Static{Vertex,Edge} is created if there are only algebraic variables.

Currently, there is very little safetly checking. First draft.