A complete working example of the kura_plasticity
Now, an edge can have any number of 'expression' objects, which are not added to the ODE/DAE
state vector.

No checks on performance have been performed.

The current state of this branch is not compatible with other types of
systems - nd_ODE_static, etc.. have not been updated to suit the format.
But any nd_ODE_ODE example can be made to work, just note that the edge
function now has an additional argument for the array that holds the
edge expression objects.

See examples/kuramoto_plasticity.jl

-L